feat!: Velcro component yields hash instead of values
Velcro component now yields a hash.

Before:

`<Velcro as |x y z|></Velcro>`

After:

```
<Velcro as |velcro|>
  <velcro.hook>Hook</velcro.hook>
  <velcro.loop>Hook</velcro.looop>
  {{log "velcro data" velcro.data}}
</Velcro>
```

More generally, the nomenclature for elements has been updated; the
'referenceElement' is now 'hook' and 'velcroElement' is 'loop'.

In Floating UI parlance, the 'hook' is the `reference` element and the
'loop' is the `floating` element.

Calling things 'hook' and 'loop' fits with the Velcro theme and is fun.