feat(DO-NOT-MERGE): add named blocks
An alternate way to use the Velcro component could be via named blocks:

```
<Velcro>
  <:hook></:hook>
  <:loop></:loop>
</Velcro>
```

As neat as this looks it immediately introduces a concern with layout;
there would always be a wrapping `div` which could hinder layout at
the components call site. Unless that's the markup you want but then
there would be no way to style the hook's div without spreading the
attributes onto the 'hook' element which is just confusing; writing
styles on `<Velcro class="foo">` would imply there is a wrapping
element.

This syntax would be better applied to components composed from Ember
Velcro with established styles for the UI's look and feel. Probably.